trim the chart stories to the two that earn their keep
Drop the palette-comparison and deep-slot views — they existed to choose
between candidate palettes, and that choice is made. What's left is one
colour-vision grid and the legend interaction test. Rename the file to match
what it now covers.
